--- Assets/Scripts/GameManager - Copy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5fd75e2373c19ec4bbee025820d8e30e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
